+++ CHANGELOG.md
@@ -6,6 +6,10 @@
     - Read the default zip code and api key from the config file
     - Display usage text with --help or -h
 
+- Bug fixes
+    - Display an error if the city is not found
+    - Display an error when the api key is invalid
+    
 ## 0.0.1 (2020-11-22)
 
 - Get the current weather conditions by zip code